feat(plugins): add `show` command
